+++ RoutineSticks.h
@@ -11,8 +11,8 @@ class CRoutineSticks : public CRoutine
         static constexpr size_t   c_max_num_sticks = 32;
         static constexpr uint32_t c_min_period_sec = 20;
         static constexpr uint32_t c_max_period_sec = 40;
-        static constexpr double   c_min_width_pct  = 0.02;
-        static constexpr double   c_max_width_pct  = 0.10;
+        static constexpr float   c_min_width_pct  = 0.02;
+        static constexpr float   c_max_width_pct  = 0.10;
 
     public:
         CRoutineSticks(CPixelArray* pixels, size_t num_sticks);
